+++ packages/blade/src/components/Dropdown/DropdownOverlay.web.tsx
@@ -1,14 +1,14 @@
-import React, { useState } from 'react';
+import React from 'react';
 import throttle from 'lodash/throttle';
 import styled, { keyframes, css } from 'styled-components';
 import type { FlattenSimpleInterpolation } from 'styled-components';
-import { useFloating } from '@floating-ui/react';
+import { autoUpdate, useFloating } from '@floating-ui/react';
 import type { DropdownPosition } from './dropdownUtils';
 import { componentIds, getDropdownOverflowMiddleware } from './dropdownUtils';
 import { useDropdown } from './useDropdown';
 import { StyledDropdownOverlay } from './StyledDropdownOverlay';
 import BaseBox from '~components/Box/BaseBox';
-import { castWebType, makeMotionTime, makeSize, metaAttribute, MetaConstants } from '~utils';
+import { makeMotionTime, makeSize, metaAttribute, MetaConstants } from '~utils';
 import { useTheme } from '~components/BladeProvider';
 // Reading directly because its not possible to get theme object on top level to be used in keyframes
 import { spacing, size } from '~tokens/global';
@@ -45,6 +45,7 @@ const AnimatedOverlay = styled(StyledDropdownOverlay)<{
   transition: FlattenSimpleInterpolation;
   onAnimationEnd: () => void;
   isInBottomSheet?: boolean;
+  isOpen: boolean;
 }>(
   (props) =>
     css`
@@ -52,11 +53,12 @@ const AnimatedOverlay = styled(StyledDropdownOverlay)<{
       transform: translateY(${makeSize(props.theme.spacing[3])});
       opacity: 0;
       z-index: 99;
+      pointer-events: ${props.isOpen ? 'all' : 'none'};
     `,
 );
 
 type DropdownOverlayProps = {
-  children: React.ReactNode;
+  children: React.ReactElement;
 } & TestID;
 
 /**
@@ -65,12 +67,19 @@ type DropdownOverlayProps = {
  * Wrap your ActionList within this component
  */
 const _DropdownOverlay = ({ children, testID }: DropdownOverlayProps): JSX.Element => {
-  const { isOpen, triggererRef, hasLabelOnLeft, dropdownTriggerer, setIsOpen } = useDropdown();
+  const {
+    isOpen,
+    triggererRef,
+    hasLabelOnLeft,
+    dropdownTriggerer,
+    setIsOpen,
+    actionListItemRef,
+  } = useDropdown();
   const { theme } = useTheme();
   const bottomSheetAndDropdownGlue = useBottomSheetAndDropdownGlue();
-  const [display, setDisplay] = React.useState<'none' | 'block'>('none');
+  const [showFadeOutAnimation, setShowFadeOutAnimation] = React.useState(false);
   const [width, setWidth] = React.useState<SpacingValueType>('100%');
-  const [dropdownPosition, setDropdownPosition] = useState<DropdownPosition>({});
+  const [dropdownPosition, setDropdownPosition] = React.useState<DropdownPosition>({});
 
   const isMenu = dropdownTriggerer !== 'SelectInput';
 
@@ -82,7 +91,15 @@ const _DropdownOverlay = ({ children, testID }: DropdownOverlayProps): JSX.Eleme
     elements: {
       reference: triggererRef.current,
     },
-    middleware: [getDropdownOverflowMiddleware({ isMenu, triggererRef, setDropdownPosition })],
+    middleware: [
+      getDropdownOverflowMiddleware({
+        isMenu,
+        triggererRef,
+        actionListItemRef,
+        setDropdownPosition,
+      }),
+    ],
+    whileElementsMounted: autoUpdate,
   });
 
   const fadeIn = css`
@@ -95,11 +112,14 @@ const _DropdownOverlay = ({ children, testID }: DropdownOverlayProps): JSX.Eleme
       ${String(theme.motion.easing.entrance.revealing)};
   `;
 
+  const noAnimation = css`
+    animation: none;
+  `;
+
   React.useEffect(() => {
     if (isOpen) {
       // On Safari clicking on a non input element doesn't focuses it https://bugs.webkit.org/show_bug.cgi?id=22261
       triggererRef.current?.focus();
-      setDisplay('block');
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isOpen]);
@@ -136,11 +156,12 @@ const _DropdownOverlay = ({ children, testID }: DropdownOverlayProps): JSX.Eleme
 
   const onAnimationEnd = React.useCallback(() => {
     if (isOpen) {
-      setDisplay('block');
+      setShowFadeOutAnimation(true);
     } else {
-      setDisplay('none');
+      setShowFadeOutAnimation(false);
     }
   }, [isOpen]);
+
   const styles = React.useMemo(() => ({ opacity: isOpen ? 1 : 0 }), [isOpen]);
 
   return (
@@ -167,9 +188,9 @@ const _DropdownOverlay = ({ children, testID }: DropdownOverlayProps): JSX.Eleme
         right={dropdownPosition.right}
         bottom={dropdownPosition.bottom}
         style={styles}
-        display={castWebType(display)}
+        isOpen={isOpen}
         position="absolute"
-        transition={isOpen ? fadeIn : fadeOut}
+        transition={isOpen ? fadeIn : showFadeOutAnimation ? fadeOut : noAnimation}
         onAnimationEnd={onAnimationEnd}
         {...metaAttribute({ name: MetaConstants.DropdownOverlay, testID })}
       >