+++ packages/blade/src/components/Dropdown/docs/stories.ts
@@ -524,8 +524,6 @@ const WithControlledSelectStory = `
             <ActionList>
               <ActionListItem title="Mumbai" value="mumbai" />
               <ActionListItem title="Bangalore" value="bangalore" />
-              <ActionListItem title="Pune" value="pune" />
-              <ActionListItem title="Chennai" value="chennai" />
             </ActionList>
           </DropdownOverlay>
         </Dropdown>